+++ contracts/walrus/sources/system/events.move
@@ -11,11 +11,11 @@ use sui::event;
 
 /// Signals that a blob with meta-data has been registered.
 public struct BlobRegistered has copy, drop {
-    epoch: u64,
+    epoch: u32,
     blob_id: u256,
     size: u64,
     erasure_code_type: u8,
-    end_epoch: u64,
+    end_epoch: u32,
     deletable: bool,
     // The object id of the related `Blob` object
     object_id: ID,
@@ -23,9 +23,9 @@ public struct BlobRegistered has copy, drop {
 
 /// Signals that a blob is certified.
 public struct BlobCertified has copy, drop {
-    epoch: u64,
+    epoch: u32,
     blob_id: u256,
-    end_epoch: u64,
+    end_epoch: u32,
     deletable: bool,
     // The object id of the related `Blob` object
     object_id: ID,
@@ -35,55 +35,55 @@ public struct BlobCertified has copy, drop {
 
 /// Signals that a blob has been deleted.
 public struct BlobDeleted has copy, drop {
-    epoch: u64,
+    epoch: u32,
     blob_id: u256,
-    end_epoch: u64,
+    end_epoch: u32,
     // The object ID of the related `Blob` object.
     object_id: ID,
 }
 
 /// Signals that a BlobID is invalid.
 public struct InvalidBlobID has copy, drop {
-    epoch: u64, // The epoch in which the blob ID is first registered as invalid
+    epoch: u32, // The epoch in which the blob ID is first registered as invalid
     blob_id: u256,
 }
 
 /// Signals that epoch `epoch` has started and the epoch change is in progress.
 public struct EpochChangeStart has copy, drop {
-    epoch: u64,
+    epoch: u32,
 }
 
 /// Signals that a set of storage nodes holding at least 2f+1 shards have finished the epoch
 /// change, i.e., received all of their assigned shards.
 public struct EpochChangeDone has copy, drop {
-    epoch: u64,
+    epoch: u32,
 }
 
 /// Signals that a node has received the specified shards for the new epoch.
 public struct ShardsReceived has copy, drop {
-    epoch: u64,
+    epoch: u32,
     shards: vector<u16>,
 }
 
 /// Signals that the committee and the system parameters for `next_epoch` have been selected.
 public struct EpochParametersSelected has copy, drop {
-    next_epoch: u64,
+    next_epoch: u32,
 }
 
 /// Signals that the given shards can be recovered using the shard recovery endpoint.
 public struct ShardRecoveryStart has copy, drop {
-    epoch: u64,
+    epoch: u32,
     shards: vector<u16>,
 }
 
 // == Functions to emit the events from other modules ==
 
 public(package) fun emit_blob_registered(
-    epoch: u64,
+    epoch: u32,
     blob_id: u256,
     size: u64,
     erasure_code_type: u8,
-    end_epoch: u64,
+    end_epoch: u32,
     deletable: bool,
     object_id: ID,
 ) {
@@ -99,9 +99,9 @@ public(package) fun emit_blob_registered(
 }
 
 public(package) fun emit_blob_certified(
-    epoch: u64,
+    epoch: u32,
     blob_id: u256,
-    end_epoch: u64,
+    end_epoch: u32,
     deletable: bool,
     object_id: ID,
     is_extension: bool,
@@ -109,30 +109,30 @@ public(package) fun emit_blob_certified(
     event::emit(BlobCertified { epoch, blob_id, end_epoch, deletable, object_id, is_extension });
 }
 
-public(package) fun emit_invalid_blob_id(epoch: u64, blob_id: u256) {
+public(package) fun emit_invalid_blob_id(epoch: u32, blob_id: u256) {
     event::emit(InvalidBlobID { epoch, blob_id });
 }
 
-public(package) fun emit_blob_deleted(epoch: u64, blob_id: u256, end_epoch: u64, object_id: ID) {
+public(package) fun emit_blob_deleted(epoch: u32, blob_id: u256, end_epoch: u32, object_id: ID) {
     event::emit(BlobDeleted { epoch, blob_id, end_epoch, object_id });
 }
 
-public(package) fun emit_epoch_change_start(epoch: u64) {
+public(package) fun emit_epoch_change_start(epoch: u32) {
     event::emit(EpochChangeStart { epoch })
 }
 
-public(package) fun emit_epoch_change_done(epoch: u64) {
+public(package) fun emit_epoch_change_done(epoch: u32) {
     event::emit(EpochChangeDone { epoch })
 }
 
-public(package) fun emit_shards_received(epoch: u64, shards: vector<u16>) {
+public(package) fun emit_shards_received(epoch: u32, shards: vector<u16>) {
     event::emit(ShardsReceived { epoch, shards })
 }
 
-public(package) fun emit_epoch_parameters_selected(next_epoch: u64) {
+public(package) fun emit_epoch_parameters_selected(next_epoch: u32) {
     event::emit(EpochParametersSelected { next_epoch })
 }
 
-public(package) fun emit_shard_recovery_start(epoch: u64, shards: vector<u16>) {
+public(package) fun emit_shard_recovery_start(epoch: u32, shards: vector<u16>) {
     event::emit(ShardRecoveryStart { epoch, shards })
 }
